+++ test/radar/reporter/logger_reporter_test.rb
@@ -10,10 +10,6 @@ class LoggerReporterTest < Test::Unit::TestCase
       assert_raises(ArgumentError) { @klass.new.report(create_exception_event) }
     end
 
-    should "raise an argument error if an invalid logger is given" do
-      assert_raises(ArgumentError) { @klass.new(:log_object => 7).report(create_exception_event) }
-    end
-
     should "raise an argument error if the logger doesn't respond to the log level" do
       assert_raises(ArgumentError) { @klass.new(:log_object => Logger.new(nil), :log_level => :bananas).report(create_exception_event) }
     end